Update ps 23.12.15

+++ circle_3/philosophers/philo_bonus/srcs/chef_bonus_2.c
@@ -6,7 +6,7 @@
 /*   By: singeonho <singeonho@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/09 00:57:47 by singeonho         #+#    #+#             */
-/*   Updated: 2023/11/13 20:40:19 by singeonho        ###   ########.fr       */
+/*   Updated: 2023/12/15 15:56:39 by singeonho        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,17 +34,26 @@ void	chef_start_serving(t_chef *chef)
 {
 	pid_t	pid;
 	int		idx;
-	t_int64	tserve;
+	int		status;
 
 	idx = -1;
-	tserve = get_time();
 	while (++idx < chef->number_of_philo)
 	{
-		pid = serve_idx_philo(&(chef->philos[idx]), tserve);
+		pid = serve_idx_philo(&(chef->philos[idx]), get_time());
 		chef->pids[idx] = pid;
 	}
-	waitpid(0, NULL, 0);
-	idx = -1;
-	while (++idx < chef->number_of_philo)
-		kill(chef->pids[idx], SIGTERM);
+	idx = 0;
+	while (idx < chef->number_of_philo)
+	{
+		waitpid(-1, &status, 0);
+		if (status == 0)
+			idx++;
+		else
+		{
+			idx = -1;
+			while (++idx < chef->number_of_philo)
+				kill(chef->pids[idx], SIGTERM);
+			return ;
+		}
+	}
 }
